# openapi.yaml
openapi: 3.0.3
info:
  title: USBI Platform API
  description: |
    API institucional del sistema USBI. Todos los errores usan RFC 7807 (Problem Details).
    Todos los endpoints autenticados requieren JWT Bearer.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.usbi.edu.mx/api/v1
    description: Servidor de producción (On-Premise, TLS 1.2+)
security:
  - bearerAuth: []
paths:
...
  /auth/tutor-consent:
    post:
      summary: Solicitud de consentimiento de tutor (paso 1 del doble opt-in)
      operationId: submitTutorConsent
      security: []
      description: |
        Inicia el doble opt-in del tutor. Guarda los datos del tutor como una
        solicitud PENDIENTE con un token de un solo uso (24h por defecto) y envía
        un enlace de verificación al correo del tutor. La cuenta del menor NO se
        activa aquí: la activación ocurre sólo cuando el tutor abre el enlace
        (GET /auth/tutor-consent/verify), momento en que se registran la IP, la
        fecha y la hora del clic como evidencia legal del consentimiento.

        La respuesta es intencionalmente uniforme (202) exista o no la cuenta y
        esté o no pendiente, para no filtrar la existencia de cuentas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorConsentRequestDTO'
      responses:
        '202':
          description: |
            Solicitud aceptada. Si la cuenta requiere consentimiento, se envió un
            correo de verificación al tutor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending_verification]
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Error de validación de los datos del tutor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '502':
          description: No se pudo enviar el correo de verificación al tutor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/tutor-consent/verify:
    get:
      summary: Verificación del consentimiento de tutor (paso 2 del doble opt-in)
      operationId: verifyTutorConsent
      security: []
      description: |
        Endpoint destinado al navegador del tutor (enlace del correo). Valida el
        token, registra el clic (IP, fecha y hora) como evidencia legal, crea el
        registro de consentimiento con sello HMAC y activa la cuenta del menor.
        Responde con una página HTML (no JSON) porque el tutor la abre desde su
        cliente de correo.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Token opaco de un solo uso entregado en el correo del tutor.
      responses:
        '200':
          description: Consentimiento confirmado; cuenta activada. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Token ausente o inválido. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '409':
          description: El token ya había sido utilizado. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '410':
          description: El token caducó. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Error interno al procesar la verificación. Página HTML.
          content:
            text/html:
              schema:
                type: string
...
  /arco:
    post:
      summary: Solicitud de Derechos ARCO (Acceso, Rectificación, Cancelación, Oposición)
      operationId: submitArcoRequest
      description: |
        Registra una solicitud ARCO. Si el tipo es 'cancelacion', el backend
        marcará wipe_local_data=true en la tabla devices para el usuario,
        que será inyectado en la siguiente respuesta de sync.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArcoRequestDTO'
      responses:
        '201':
          description: Solicitud ARCO registrada con evidence_hash.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArcoResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /arco/pending:
    get:
      summary: Listar solicitudes ARCO pendientes
      operationId: listPendingArcoRequests
      description: Sólo admin/director. Devuelve metadatos técnicos sin PII.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Solicitudes pendientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArcoPendingListDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /arco/{request_id}/resolve:
    post:
      summary: Resolver solicitud ARCO
      operationId: resolveArcoRequest
      description: |
        Sólo admin/director. Si la solicitud aprobada es de cancelación,
        pseudonimiza PII, revoca sesiones y marca dispositivos para borrado local.
      parameters:
        - in: path
          name: request_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveArcoRequestDTO'
      responses:
        '200':
          description: Solicitud resuelta.
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /devices:
    get:
      summary: Listar dispositivos registrados del usuario autenticado
      operationId: listDevices
      responses:
        '200':
          description: Dispositivos activos del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponseDTO'
    post:
      summary: Registrar dispositivo para sincronización offline
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequestDTO'
      responses:
        '201':
          description: Dispositivo registrado. Usar `id` como device_id en sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDTO'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT firmado con HS256. Incluye user_id, role y token_version.
  schemas:
...
    TutorConsentRequestDTO:
      type: object
      required: [user_id, tutor_name, tutor_email, privacy_notice_version]
      additionalProperties: false
      properties:
        user_id:
          type: string
          format: uuid
        tutor_name:
          type: string
        tutor_email:
          type: string
          format: email
        privacy_notice_version:
          type: string

    # ── User ──────────────────────────────────────────────────────────────────
...
    RegisterDeviceRequestDTO:
      type: object
      required: [device_label, platform]
      additionalProperties: false
      properties:
        device_label:
          type: string
        platform:
          type: string
          enum: [web, tauri]
    DeviceDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_label:
          type: string
        platform:
          type: string
          enum: [web, tauri]
        registered_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        wipe_local_data:
          type: boolean
    DeviceListResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DeviceDTO'

    # ── ARCO ──────────────────────────────────────────────────────────────────
    ArcoRequestDTO:
      type: object
      required: [request_type]
      additionalProperties: false
      properties:
        request_type:
          type: string
          enum: [acceso, rectificacion, cancelacion, oposicion]
        details:
          type: string
          maxLength: 1000
    ArcoResponseDTO:
      type: object
      required: [request_id, status, message]
      properties:
        request_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
        message:
          type: string
    ArcoPendingItemDTO:
      type: object
      required: [id, requester_type, request_type, status, received_at]
      properties:
        id:
          type: string
          format: uuid
        requester_type:
          type: string
        request_type:
          type: string
          enum: [acceso, rectificacion, cancelacion, oposicion]
        status:
          type: string
          enum: [pending]
        received_at:
          type: string
          format: date-time
    ArcoPendingListDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArcoPendingItemDTO'
    ResolveArcoRequestDTO:
      type: object
      required: [approved, response_summary]
      additionalProperties: false
      properties:
        approved:
          type: boolean
        response_summary:
          type: string
          maxLength: 1000
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.usbi.edu.mx/errors/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: El campo email es requerido.
        instance:
          type: string
          format: uri
          example: /api/v1/auth/login
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: No autenticado o JWT expirado/invalidado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Sin permisos para este recurso
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'